requirements.txt
fireworks-ai==0.15.6
gradio==5.1.0
instructor==1.6.3
